Refactor code to make it easier to: 1. Add new resource. 2. Add log for new resource. 3. Add new rule for existing resource.
